requirements.txt
certifi==2017.4.17
chardet==3.0.3
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
openapi-codec==1.3.1
PyMySQL==0.7.11
pytz==2017.2
requests==2.16.5
simplejson==3.10.0
Unipath==1.1
uritemplate==3.0.0
urllib3==1.21.1
